bugfix(fivetran/unity): move UnityCatalogConnectionConfig import to avoid circular deps with ge_profiler
